+++ prowler/providers/m365/services/defenderendpoint/defenderendpoint_privileged_user_exposed_credentials/defenderendpoint_privileged_user_exposed_credentials.metadata.json
@@ -1,10 +1,10 @@
 {
   "Provider": "m365",
-  "CheckID": "defender_endpoint_privileged_user_exposed_credentials",
+  "CheckID": "defenderendpoint_privileged_user_exposed_credentials",
   "CheckTitle": "Ensure privileged users do not have credentials exposed on vulnerable endpoints",
   "CheckType": [],
-  "ServiceName": "defender",
-  "SubServiceName": "endpoint",
+  "ServiceName": "defenderendpoint",
+  "SubServiceName": "",
   "ResourceIdTemplate": "",
   "Severity": "critical",
   "ResourceType": "Exposure Management",
